rename presenter to format
