+++ chronicle/release/format/format.go
@@ -1,4 +1,4 @@
-package presenter
+package format
 
 import "strings"
