+++ chronicle/release/format/markdown/test-fixtures/snapshot/TestMarkdownPresenter_Present.golden
